auto (pause.c) 52 1010

+++ pause.c
@@ -27,31 +27,39 @@ int main(void) {
 	
   gtk_init(NULL,NULL);
 
-  GtkBuilder* builder = gtk_builder_new_from_string(gladeFile,gladeFile_length);
+  GtkBuilder* builder = gtk_builder_new_from_string((gchar*)gladeFile,gladeFile_length);
   GtkWidget* top = GTK_WIDGET(gtk_builder_get_object(builder,"top"));
 	GtkImage* image = GTK_IMAGE(gtk_builder_get_object(builder,"image"));
   gtk_window_stick(GTK_WINDOW(top));
   gtk_window_set_keep_above(GTK_WINDOW(top),TRUE);
 
 	bool stopped = false;
-	static void toggle(GtkWidget* top, GdkEventButton* e, gpointer udata) {
+	gboolean toggle(GtkWidget* top, GdkEventButton* e, gpointer udata) {
 		if(e->state & GDK_CONTROL_MASK) {
 			gtk_window_begin_move_drag(GTK_WINDOW(top), e->button, e->x_root, e->y_root, e->time);
-			return;
+			return FALSE;
 		}
+		if(e->state & GDK_SHIFT_MASK) {
+			gtk_main_quit();
+			return TRUE;
+		}
+		printf("uh %d\n",stopped);
 		if(stopped) {
 			kill(pid,SIGCONT);
 			stopped = false;
-			gtk_image_set_from_stock(image, "gtk-stop", GTK_ICON_SIZE_BUTTON);
+			gtk_image_set_from_icon_name(image, "gtk-stop", GTK_ICON_SIZE_LARGE_TOOLBAR);
+			gtk_widget_set_tooltip_text(top, "Pause");
 		} else {
 			//printf("stop %d\n",pid);
 			kill(pid, SIGSTOP);
 			stopped = true;
-			gtk_image_set_from_stock(image, "gtk-media-play", GTK_ICON_SIZE_BUTTON);
+			gtk_image_set_from_icon_name(image, "gtk-media-play", GTK_ICON_SIZE_LARGE_TOOLBAR);
+			gtk_widget_set_tooltip_text(top, "Play");
 		}
+		return TRUE;
 	}
 
-  g_signal_connect(G_OBJECT(top),"button-release-event",G_CALLBACK(toggle),NULL);
+  g_signal_connect(G_OBJECT(top),"button-release-event",G_CALLBACK(toggle),top);
 	
   gtk_widget_show_all(top);
   gtk_main();
